+++ ElvUI_EltreumUI/Localization/zhTW.lua
@@ -1601,3 +1601,4 @@ L["Reinstall Thin Layout"] = true
 L["Reset layout to Eltruism Thin"] = true
 L["Reverse Focus"] = true
 L["Reverse Focus Gradient Health"] = true
+L["Hide Quests during Mythic Plus"] = true